fix: ts without outDir

+++ packages/router/tsbin/sfa-router-build.ts
@@ -19,7 +19,9 @@ if (Config.tsconfig) {
   } else {
     console.log(tscResult.stdout);
   }
-  deleteFile(outDir, ".d.ts");
+  if (outDir) {
+    deleteFile(outDir, ".d.ts");
+  }
 
   if (outDir) {
     if (Config.default.ts && Config.default.ts.static) {
